requirements.txt
scrapy
xlrd
mysql-connector-python
python-dotenv